Remove soldermask layers over front panel.
We will go for a black panel in the first instance, since it seems the gerber viewer doesn't render the silkscreen when it's not on top of solder mask. For JLC bare aluminium panels we should cut the silkscreen layer out of the soldermask rectangle(s) (eg, difference in Inkscape) so that our solder mask doesn't cover the silkscreen.